# Refactor AppendError to use strings.Contains